requirements.txt
asgiref==3.8.1
Django==3.0.3
django-browser-reload
python-decouple==3.8
sqlparse==0.5.0
djangorestframework==3.11.1
drf-yasg==1.21.7
factory_boy==3.0.1